imagine casting an int to an int

+++ tj_adapt_vqe/experiments.py
@@ -50,7 +50,7 @@ def train_function(
 
     starting_ansatz: list[Ansatz] = [HartreeFockAnsatz()]
 
-    max_adapt_iter = int(2 * n_qubits)
+    max_adapt_iter = 2 * n_qubits
 
     vqe = ADAPTVQE(
         molecule,
